Move XamlApplication from example to package
